refactor: isolate reusable Flow DiT graph runner
